Add: Task 13 - Testing

+++ Task8/src/main/java/org/andersen/service/UserService.java
@@ -1,6 +1,7 @@
 package org.andersen.service;
 
 import lombok.RequiredArgsConstructor;
+import org.andersen.exception.EmptyUserNameException;
 import org.andersen.exception.UserActivationIsDisabledException;
 import org.andersen.exception.UserNotFoundException;
 import org.andersen.model.user.User;
@@ -14,11 +15,17 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
     private final UserRepository userRepository;
 
-    @Value("${app.conditionalBean.available}")
-    private boolean isActivationEnabled;
+    @Value("${app.activate-user.enabled}")
+    private final boolean isActivationEnabled;
 
     @Transactional
-    public User addUser(User user){
+    public User addUser(User user) {
+        if (user == null) {
+            throw new NullPointerException();
+        }
+        if (user.getName().isBlank()) {
+            throw new EmptyUserNameException();
+        }
         return userRepository.save(user);
     }
 
@@ -29,10 +36,14 @@ public class UserService {
 
     @Transactional
     public void deleteUser(Long id){
+        if (id == null) {
+            throw new NullPointerException("User id is null");
+        }
         User user = userRepository.findById(id).orElseThrow(()->new UserNotFoundException());
         userRepository.delete(user);
     }
 
+    @Transactional
     public void activateUser(long id) throws UserActivationIsDisabledException {
         if (isActivationEnabled) {
             userRepository.updateUserStatus(id);
